better separation of codes
